requirements.txt
flask==2.1.0
requests==2.27.1
python-dotenv==0.20.0
gunicorn==20.1.0